fix(guardrails): migrate hashArgs and memory fingerprint off nested-key-filtering replacer
The same JSON.stringify(value, sortedKeysArray) property-list replacer
anti-pattern that #2060's spiral-hash fix addressed also existed in
file-authority.hashArgs (guardrails circuit-breaker repetition detection +
provider-failure fingerprint) and redaction.computeMemoryCohortFingerprint
(memory cohort config fingerprint). The replacer filters keys at every object
depth, so nested-args tools (e.g. todowrite with a todos array) collapsed
distinct nested content to the same hash — the exact collision class that
breaks repetition detection.

Extract stableCanonicalStringify to src/utils/stable-stringify.ts as a shared
recursive canonical serializer (sorts keys at every depth, no filtering). All
three call sites (hashArgsForSpiral, hashArgs, computeMemoryCohortFingerprint)
now use it; adversarial-detector's local copy is removed in favor of the
shared import.

The redaction.ts migration is byte-identical for the current flat input shape
(verified empirically), so existing persisted fingerprints remain valid. The
hashArgs change only affects nested args (the fix); flat args are unchanged and
the value is never persisted across versions (recentToolCalls is reset on
snapshot rehydration).

Independently reviewed (Kimi K2.7) and critic-gated (Kimi K3); the reviewer
caught a third sibling site (redaction.ts) that the initial fix missed.
