+++ docs/releases/pending/2059-2060-write-target-and-spiral.md
@@ -2,8 +2,10 @@
 
 ## What
 
-Two independent guardrail-hardening fixes that prevent legitimate agent tool
-calls from being blocked by false-positive verifications.
+Three independent guardrail-hardening fixes that prevent legitimate agent tool
+calls from being blocked by false-positive verifications, plus a shared stable
+serialization helper that eliminates a class of nested-key-filtering bugs
+across the codebase.
 
 - **#2059 — `write-target-resolver`:** the patch resolver now recognizes the
   common payload aliases `patchText`, `patch_text`, `patchPayload`, `text`, and
@@ -13,6 +15,11 @@ calls from being blocked by false-positive verifications.
 - **#2060 — `adversarial-detector`:** the debugging-spiral detector now hashes
   the complete tool-call arguments with a fixed-length, key-sorted hash instead
   of truncating the stringified args at 100 characters.
+- **`file-authority.hashArgs` + `redaction.computeMemoryCohortFingerprint`:**
+  migrated off the same `JSON.stringify(value, sortedKeysArray)` property-list
+  replacer anti-pattern that #2060's fix addressed, using a shared
+  `stableCanonicalStringify` helper so all three call sites use one correct
+  implementation.
 
 ## Why
 
@@ -56,12 +63,40 @@ message.
 - `recordToolCall()` now uses a new `hashArgsForSpiral()` helper that produces
   a fixed-length hash via `bunHash()` (the existing cross-runtime shim: xxHash64
   on Bun, djb2 on Node). Object keys are sorted recursively at every depth via
-  a dedicated `stableCanonicalStringify` helper (not a `JSON.stringify`
+  the shared `stableCanonicalStringify` helper (not a `JSON.stringify`
   property-list replacer, which would silently drop nested keys and
   re-introduce collisions for nested-args tools like `todowrite`). Strings
   above 64 chars are hashed too, keeping per-entry memory bounded (net decrease
   vs. the old 100-char ceiling). `SPIRAL_THRESHOLD` (5) is unchanged — the fix
   is in the hash, not the threshold.
+
+### `src/utils/stable-stringify.ts` (new)
+- Exports `stableCanonicalStringify`, a recursive canonical JSON serializer
+  that sorts object keys at every depth without filtering (unlike a
+  `JSON.stringify` property-list replacer, which filters keys at nested
+  depths). Shared by `hashArgsForSpiral`, `hashArgs`, and
+  `computeMemoryCohortFingerprint` so all three use one correct implementation.
+
+### `src/hooks/guardrails/file-authority.ts`
+- `hashArgs` (used by the guardrails circuit-breaker repetition detector and
+  the provider-failure fingerprint) migrated from
+  `JSON.stringify(args, sortedKeys)` to `stableCanonicalStringify(args)`. This
+  fixes the same nested-key-filtering bug class as #2060 for the circuit
+  breaker: nested-args tools (e.g. `todowrite` with a `todos` array) previously
+  collapsed distinct nested content to the same hash, so a genuine repetition
+  loop on a nested-args tool could be missed. The returned `number` is
+  unchanged for flat args (the common case) and only differs for nested args
+  (the fix). The value is never persisted across versions
+  (`recentToolCalls` is reset on snapshot rehydration), so there is no
+  cross-version comparison risk.
+
+### `src/memory/redaction.ts`
+- `computeMemoryCohortFingerprint` migrated to `stableCanonicalStringify` for
+  consistency and to future-proof against a nested field ever being added to
+  `MemoryCohortFingerprintInput`. For the current flat input shape the output
+  is byte-identical to the prior `JSON.stringify(input, sortedKeys)` — verified
+  empirically — so existing persisted fingerprints in
+  `memory-cohort-config.json` remain valid.
 
 ## Test plan
 - New `write-target-resolver.test.ts` cases: each alias resolves a unified
@@ -73,6 +108,12 @@ message.
   args spiral (correctness); string vs object args do not collide; nested-args
   tools with different nested content do not spiral; nested-args with reordered
   keys at any depth still spiral.
+- New `guardrails.test.ts` cases for `hashArgs`: nested args with different
+  content produce different hashes (no nested-key filtering); nested args with
+  reordered keys at any depth produce the same hash. Both verified to fail
+  against the old replacer-array code.
+- `cohort-config-fingerprint.test.ts` passes unchanged (redaction.ts migration
+  is a no-op for the current flat input — verified byte-identical).
 - All existing write-target, scope-guard, guardrails, and adversarial-detector
   tests pass unchanged.
